requirements.txt

python-telegram-bot==13.12
openai
python-dotenv
